requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
pydantic>=2.0
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
python-multipart
